+++ src/GenCP_util.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, Bart Swinnen
+ * Copyright (c) 2020-2021, Bart Swinnen
  * All rights reserved.
  *
  * BSD 2-Clause License (see included LICENSE file)
@@ -8,7 +8,8 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 
 #include "GenCP_def.h"
 
-uint16_t crc16(uint16_t * buffer, uint32_t buffer_size_bytes, gcp_endianness_e endianness);
+uint16_t crc16(uint16_t * buffer, size_t buffer_size_bytes, gcp_endianness_e endianness);
